+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.12.0
+
+- Add "results" word to search results: No results, 1 result, 2 results (with localization support).
+- Make search results count easier to see: Increase font size and brightness.
+
 ## 0.11.0 - August 9, 2025
 
 - Increase font size for lexicon.